Default topics used to send and receive messages from the JSON adapter are now '$fabric/$adapters/$mqtt/$in' and '$fabric/$adapters/$mqtt/$out'.
The change is to remove the insertion point ({0}) for the Fabric node
name, which is only required if multiple Edgware nodes share a broker.

This change has been made to simplify client code, by not requiring the
client to know the Edgware node name, just the broker IP name/address
and port.