Support added for new "$save" directive, and associated variable substitution.
A script can include the "$save <correlation-id>" directive to wait for
and record a message with a specified correlation ID. Subsequently,
fields from the saved message can be substituted into messages sent to
Edgware.

The syntax is best illustrated with an example:

# Wait for a message with a correlation ID of 12345 and save it
$save 12345

# Insert a field from 12345 into a message:
$send
{
	"op" : "publish",
	"output-feed" : "replay_platform/temperature/untyped_output_feed",
	"msg" : "%%12345:temperature%%"
}

The reference "%%12345:temperature%%" will be replaced with the field
"temperature" from the saved message with correlation ID "12345".